Support enabling/disabling restore-backup process by parameter

+++ scripts/deploy/backup-environment.sh
@@ -11,7 +11,10 @@
 PARAMS_COUNT=${#}
 
 function copy() {
-	rsync -av --delete ${1} ${2}
+	if [ ${3} == true ];
+	then
+		rsync -av --delete ${1} ${2}
+	fi
 }
 
 function backupDatabase() {
@@ -23,9 +26,9 @@ function restoreDatabase() {
 }
 
 function usage() {
-	if [ ${PARAMS_COUNT} -ne 5 ]
+	if [ ${PARAMS_COUNT} -ne 6 ]
 	then
-		echo "Please use this script with five parameters: host, environment-name, operation {restore | backup}, db-user and db-password. Actually ${PARAMS_COUNT} parameters have been sent."
+		echo "Please use this script with six parameters: host, environment-name, operation {restore | backup}, db-user, db-password and backup/restore enabled {1 | 0}. Actually ${PARAMS_COUNT} parameters have been sent."
 		exit 1
 	fi
 }
@@ -34,10 +37,11 @@ usage
 
 # 
 # ${1} : hostname where files will be copied from/to
-# ${2} : server name {master | master_demo | master_erpl | master_so | master_pre_alloy2}
+# ${2} : server name {master | master_demo | master_erpl | master_so | master_pre_alloy2 | master_lrdcom}
 # ${3} : type of operation {restore | backup}
 # ${4} : database user
 # ${5} : database password 
+# ${6} : backup/restore enabled or disabled {true | false}
 #
 
 HOST=${1}
@@ -46,6 +50,8 @@ OPERATION_TYPE=${3}
 
 DB_USER=${4}
 DB_PASSWORD=${5}
+
+BACKUP_RESTORE_ENABLED=${6}
 
 BACKUP_PREFIX="/backup"
 
@@ -61,10 +67,10 @@ DB_BACKUP_ENVIRONMENT_FILENAME="${DIR_BACKUP_HOME_ENVIRONMENT}/liferay_${ENVIRON
 if [ "${OPERATION_TYPE}" == "restore" ]
 then
 	# restore portal application server
-	copy ${DIR_BACKUP_APP_ENVIRONMENT}/${ENVIRONMENT} ${DIR_APP_ENVIRONMENT}/
+	copy ${DIR_BACKUP_APP_ENVIRONMENT}/${ENVIRONMENT} ${DIR_APP_ENVIRONMENT}/ ${BACKUP_RESTORE_ENABLED}
 
 	# restore portal home
-	copy ${DIR_BACKUP_HOME_ENVIRONMENT}/${ENVIRONMENT} ${DIR_HOME_ENVIRONMENT}/
+	copy ${DIR_BACKUP_HOME_ENVIRONMENT}/${ENVIRONMENT} ${DIR_HOME_ENVIRONMENT}/ ${BACKUP_RESTORE_ENABLED}
 
 	# restore database
 	restoreDatabase ${HOST} ${DB_USER} ${DB_PASSWORD} ${ENVIRONMENT} ${DB_BACKUP_ENVIRONMENT_FILENAME}
@@ -72,10 +78,10 @@ else
 	if [ "${OPERATION_TYPE}" == "backup" ]
 	then
 		# backup portal application server
-		copy ${DIR_APP_ENVIRONMENT}/${ENVIRONMENT} ${DIR_BACKUP_APP_ENVIRONMENT}/
+		copy ${DIR_APP_ENVIRONMENT}/${ENVIRONMENT} ${DIR_BACKUP_APP_ENVIRONMENT}/ ${BACKUP_RESTORE_ENABLED}
 
 		# backup portal home
-		copy ${DIR_HOME_ENVIRONMENT}/${ENVIRONMENT} ${DIR_BACKUP_HOME_ENVIRONMENT}/
+		copy ${DIR_HOME_ENVIRONMENT}/${ENVIRONMENT} ${DIR_BACKUP_HOME_ENVIRONMENT}/ ${BACKUP_RESTORE_ENABLED}
 
 		# backup database
 		backupDatabase ${HOST} ${DB_USER} ${DB_PASSWORD} ${ENVIRONMENT} ${DB_BACKUP_ENVIRONMENT_FILENAME}
